Replaced pathfinding code with a heavily modified code from  https://github.com/ClintFMullins/JumpPointSearch-Java
This is only temporary and will be replaced later when the rest of the basic game is working and I can get my own branch for pathfinding. I also make the mouse cursor point in the right spot, and added a pathfinding debugger that can be used with the T key.