Entities now have an arrayList of abilities. Abilities can either be passive, changed/activated with a single click either on the button in-game or on the keyboard, and active ones require you to press the icon/key then click on a location or unit target. Active abilities will now change the cursor icon and disable cursor selection.
Added a test "Blink" ability which can be activated by pressing B with a unit selected, then clicking where you want to blink to. This is a short-range teleport which will only work if the target location is walkable, and if the goal is farther away than the range, it will select the nearest tile within range to test.